PROD-2298 #comment move the work-provider to within the work tool #time 1.5h
