+++ src-ts/index.ts
@@ -2,7 +2,6 @@ export { default as AppNextGen } from './App'
 export { EnvironmentConfig } from './config'
 export {
      Breadcrumb,
-     ChallengeMetadataName,
      ContactSupportModal,
      logInitialize,
      OrderContractModal,
@@ -12,9 +11,6 @@ export {
      RouteProvider,
      TabsNavbar,
      TermsModal,
-     workContext,
-     WorkStatusItem,
-     WorkType,
 } from './lib'
-export * from './utils'
 export * from './tools'
+export * from './utils'
