+++ src-ts/lib/functions/index.ts
@@ -5,7 +5,6 @@ export {
 } from './authentication-functions'
 export * from './component-visible-functions'
 export * from './logging-functions'
-export * from './message-functions'
 export * from './text-format-functions'
 export * from './user-functions'
 export * from './xhr-functions'
