+++ src-ts/tools/index.ts
@@ -1,10 +1,15 @@
 export { default as ToolsRoutes } from './tools.routes'
 export {
+    ChallengeMetadataName,
     selfServiceRootRoute as routeRootLoggedOut,
-    WorkFeedback,
+    workContext,
     WorkDetailDetails,
     WorkDetailHeader,
     WorkDetailSummary,
     WorkDetailSolutions,
+    workFactoryMapFormData,
+    WorkFeedback,
     workRootRoute as routeRootLoggedIn,
+    WorkStatusItem,
+    WorkType,
 } from './work'
