+++ src-ts/tools/work/index.ts
@@ -2,4 +2,5 @@ export * from './work-detail-details'
 export * from './work-detail-header'
 export * from './work-detail-solutions'
 export * from './work-detail-summary'
+export * from './work-lib'
 export * from './work.routes'
