+++ src-ts/tools/work/work-detail-solutions/work-solutions-list/work-solutions-list-item/WorkSolutionsListItem.tsx
@@ -1,6 +1,7 @@
 import { FC } from 'react'
 
-import { Button, WorkSolution } from '../../../../../lib'
+import { Button } from '../../../../../lib'
+import { WorkSolution } from '../../../work-lib'
 
 import styles from './WorkSolutionsListItem.module.scss'
