+++ src-ts/tools/work/work-detail-summary/WorkDetailProgress/WorkDetailProgressItem/WorkDetailProgressItem.tsx
@@ -1,7 +1,8 @@
 import classNames from 'classnames'
 import { FC } from 'react'
 
-import { textFormatDateLocaleShortString, WorkProgressStep } from '../../../../../lib'
+import { textFormatDateLocaleShortString } from '../../../../../lib'
+import { WorkProgressStep } from '../../../work-lib'
 
 import styles from './WorkDetailProgressItem.module.scss'
