+++ src-ts/tools/work/work-lib/work-provider/work-functions/work-factory/work-type-category.enum.ts
