+++ src-ts/tools/work/work-lib/work-provider/work-functions/work-store/challenge-phase.ts
