+++ src-ts/tools/work/work-lib/work-status-item/WorkStatusItem.module.scss
@@ -1,4 +1,4 @@
-@import '../styles/variables';
+@import '../../../../lib/styles/variables';
 
 .status-container {
     display: flex;
