+++ src-ts/tools/work/work-table/work-status-renderer/WorkStatusRenderer.tsx
@@ -1,5 +1,4 @@
-
-import { Work, WorkStatusItem } from '../../../../lib'
+import { Work, WorkStatusItem } from '../../work-lib'
 
 function WorkStatusRenderer(work: Work): JSX.Element {
     return (
