+++ src-ts/tools/work/work-table/work-table-title-renderer/WorkTableTitleRenderer.tsx
@@ -1,12 +1,11 @@
 import { FC, SVGProps } from 'react'
 
 import {
-    Work,
-    WorkTypeCategory,
     WorkTypeCategoryDataIcon,
     WorkTypeCategoryDesignIcon,
     WorkTypeCategoryUnknownIcon,
 } from '../../../../lib'
+import { Work, WorkTypeCategory } from '../../work-lib'
 
 import styles from './WorkTableTitleRenderer.module.scss'
